X11 and macOS: the next frame due an interval after this one began, not after it ended
The paint throttle waited 1000/max_fps ms after each paint, so a frame
took the paint and the interval. Now the next frame is due an interval
after the paint began (async_io::Timer::at), in microseconds rather than
whole milliseconds, as Windows' pacer does. Lingmo (llvmpipe, a VM's
software renderer) under continuous output: 21 -> 26 paints a second.

+++ window/src/os/windows/window.rs
@@ -1652,7 +1652,7 @@ unsafe fn wm_paint(hwnd: HWND, _msg: UINT, _wparam: WPARAM, _lparam: LPARAM) -> 
     let max_fps = inner.config.max_fps.max(1);
     pace_frame(
         window_id,
-        began + std::time::Duration::from_micros(1_000_000 / u64::from(max_fps)),
+        began + std::time::Duration::from_micros(1_000_000 / max_fps),
     );
 
     Some(0)
